fix race condition in status reporting
Apparently $BUILDKITE_BUILD_CHECKOUT_PATH is a global environment
variable as best as I can tell. If parallel jobs are running on the
buildkite server, then this can be changed while a job is running, so
the directory created in the post-checkout hook could be different
from the directory that is mounted or the directory that is checked in
the pre-exit hook.

To fix this, everything is now just a relative path instead. There are
no per-build environment variables that tell us our path, and we can't
run shell commands in the volume parameter for the buildkite-docker
plugin, but relative paths will be expanded to absolute paths when it
mounts them into the containers.

Just need to be careful that we're always in the directory we thought
we were when running these commands.